# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Riogas Gestión API",
    "description": "Catálogo de APIs de GOYA (app 3). Backend NestJS (prefijo /api) + route handlers del front Next.\nGenerado por `pnpm docs:api` con un parser estático del código; no editar a mano — las descripciones y ejemplos van en docs/api/anotaciones.yaml.",
    "version": "0.1.0"
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RevisarMatchDto": {
        "type": "object",
        "properties": {
          "accion": {
            "type": "string",
            "enum": [
              "aprobar",
              "rechazar",
              "corregir"
            ]
          },
          "usuario": {
            "type": "string",
            "minLength": 2
          },
          "calleOsmId": {
            "type": "integer"
          }
        },
        "required": [
          "accion",
          "usuario"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/calles/buscar": {
      "get": {
        "operationId": "CallesController_buscar",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "depto",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ciudad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "texto"
              ]
            },
            "description": "'texto' devuelve líneas pipe-delimited para que el VB6 parsee con Split()."
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "x-api-key (CallesApiKeyGuard)",
        "summary": "Busca calles del nomenclátor por texto y devuelve el CALID; `formato=texto` sale pipe-delimited para el VB6."
      }
    },
    "/api/calles/esquinas": {
      "get": {
        "operationId": "CallesController_esquinasDe",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "calle",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "numero",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "depto",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ciudad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "texto"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "x-api-key (CallesApiKeyGuard)",
        "summary": "Resuelve una dirección a sus dos esquinas y a coordenadas; `formato=texto` para el VB6."
      }
    },
...
    "/api/calles/matches": {
      "get": {
        "operationId": "CallesController_listarMatches",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "AUTO_CONFIRMADO",
                "A_REVISAR",
                "RECHAZADO",
                "CONFIRMADO_MANUAL"
              ]
            }
          },
          {
            "name": "take",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Cola de revisión: matches calle AS400 ↔ calle OSM, filtrados por estado.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/calles/matches/{id}": {
      "patch": {
        "operationId": "CallesController_revisar",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Resuelve un match de la cola: aprobar, rechazar o corregir a otra calle OSM.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisarMatchDto"
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/calles/matches/{id}/mapa": {
      "get": {
        "operationId": "CallesController_mapa",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Datos de mapa de un match, para dibujarlo en la pantalla de revisión.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/calles/osm/{id}": {
      "get": {
        "operationId": "CallesController_calleOsm",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Detalle de una calle OSM.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/calles/recargar": {
      "post": {
        "operationId": "CallesController_recargar",
        "tags": [
          "calles"
        ],
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "JWT (Bearer)",
        "summary": "Recarga en memoria el índice de calles del nomenclátor.",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/calles/resolver": {
      "get": {
        "operationId": "CallesController_resolver",
        "tags": [
          "calles"
        ],
        "parameters": [
          {
            "name": "nombreOsm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2
            }
          },
          {
            "name": "depto",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "localidad",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "x-goya-origen": "backend (NestJS)",
        "x-goya-archivo": "backend/src/calles/calles.controller.ts",
        "x-goya-auth": "x-api-key (CallesApiKeyGuard)",
        "summary": "Devuelve el CALID a partir del nombre OSM de la calle."
      }
    },
...
  },
...
}